+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   expiration in Safari.
 
 ### Added
+- New "Sanitize media files" page (/settings/sanitize-media) which allows to
+  remove sensitive metadata from the existing media files.
 - New frontend setting: comments.hideRepliesToBanned (false by default, can be
   adjusted on Settings / Appearance page). If it is set to true, comments that
   starts with a @-mention of banned user are replaced by placeholder ('Comment